Revert accidentally revert.

+++ cartographer/mapping/sparse_pose_graph.cc
@@ -73,8 +73,10 @@ proto::SparsePoseGraph SparsePoseGraph::ToProto() {
          old_node_index != single_trajectory_nodes.size(); ++old_node_index) {
       const auto& node = single_trajectory_nodes[old_node_index];
       if (!node.trimmed()) {
-        node_id_remapping[NodeId{trajectory_id, old_node_index}] =
-            NodeId{trajectory_id, trajectory_proto->node_size()};
+        node_id_remapping[NodeId{static_cast<int>(trajectory_id),
+                                 static_cast<int>(old_node_index)}] =
+            NodeId{static_cast<int>(trajectory_id),
+                   static_cast<int>(trajectory_proto->node_size())};
         auto* node_proto = trajectory_proto->add_node();
         node_proto->set_timestamp(
             common::ToUniversal(node.constant_data->time));
